@@ -23,10 +23,11 @@ dependencies:
   cached_network_image: ^3.4.1
   js: ^0.7.2
   url_launcher: ^6.3.1
-  youtube_player_flutter: ^8.1.0
+  youtube_player_flutter: ^9.1.1
   camera: ^0.10.5+9
   tflite_flutter: ^0.10.0  # Burada eklenen tflite_flutter paketi
   image_picker: ^1.1.2
+  google_fonts: ^6.2.1 
 
 dev_dependencies:
   flutter_test:
@@ -35,3 +36,11 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
+  assets:
+    - assets/images/domates.jpg
+    - assets/images/gül.jpg
+    - assets/images/lale.jpg
+    - assets/images/papatya.jpg
+    - assets/images/patlıcan.jpg
+    - assets/images/yeşil.jpg
+
