@@ -1,8 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 
-void main() {
-  runApp(const MyApp());
+void main() => runApp(const MyApp());
+
+class AppColors {
+  static const Color primary = Color(0xFF2C6E49);
+  static const Color secondary = Color(0xFFEAE1C8);
+  static const Color background = Color(0xFFEAE1C8);
+  static const Color darkPrimary = Color(0xFF1C4E29);
+  static const Color darkSecondary = Color(0xFFCCC5AE);
+  static const Color darkBackground = Color(0xFFEAE1C8);
 }
 
 class MyApp extends StatelessWidget {
@@ -11,18 +18,26 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Gelişmiş Fotoğraf Galerisi',
+      title: 'Fotoğraf Galerisi',
       theme: ThemeData(
-        primarySwatch: Colors.indigo,
-        brightness: Brightness.light,
+        colorScheme: ColorScheme.light(
+          primary: AppColors.primary,
+          secondary: AppColors.secondary,
+          surface: AppColors.secondary,
+          background: AppColors.background,
+        ),
+        scaffoldBackgroundColor: AppColors.background,
         useMaterial3: true,
-        fontFamily: 'Roboto',
       ),
       darkTheme: ThemeData(
-        primarySwatch: Colors.indigo,
-        brightness: Brightness.dark,
+        colorScheme: ColorScheme.dark(
+          primary: AppColors.darkPrimary,
+          secondary: AppColors.darkSecondary,
+          surface: AppColors.darkSecondary,
+          background: AppColors.darkBackground,
+        ),
+        scaffoldBackgroundColor: AppColors.darkBackground,
         useMaterial3: true,
-        fontFamily: 'Roboto',
       ),
       themeMode: ThemeMode.system,
       home: const GaleriScreen(),
@@ -34,11 +49,13 @@ class ImageCategory {
   final String name;
   final List<String> images;
   final IconData icon;
+  final bool isUserCreated;
 
   const ImageCategory({
     required this.name,
     required this.images,
     required this.icon,
+    this.isUserCreated = false,
   });
 }
 
@@ -49,15 +66,12 @@ class GaleriScreen extends StatefulWidget {
   State<GaleriScreen> createState() => _GaleriScreenState();
 }
 
-class _GaleriScreenState extends State<GaleriScreen>
-    with SingleTickerProviderStateMixin {
+class _GaleriScreenState extends State<GaleriScreen> {
   bool _isListView = false;
-  int _currentCrossAxisCount = 3;
-  late TabController _tabController;
+  int _gridColumns = 3;
   bool _isSearching = false;
   final TextEditingController _searchController = TextEditingController();
 
-  // Image categories
   final List<ImageCategory> _categories = [
     ImageCategory(
       name: 'Doğa',
@@ -68,77 +82,22 @@ class _GaleriScreenState extends State<GaleriScreen>
         "https://picsum.photos/id/12/800",
         "https://picsum.photos/id/13/800",
         "https://picsum.photos/id/14/800",
-        "https://picsum.photos/id/15/800",
-      ],
-    ),
-    ImageCategory(
-      name: 'Şehir',
-      icon: Icons.location_city,
-      images: [
-        "https://picsum.photos/id/16/800",
-        "https://picsum.photos/id/17/800",
-        "https://picsum.photos/id/18/800",
-        "https://picsum.photos/id/19/800",
-        "https://picsum.photos/id/20/800",
-      ],
-    ),
-    ImageCategory(
-      name: 'Hayvanlar',
-      icon: Icons.pets,
-      images: [
-        "https://picsum.photos/id/21/800",
-        "https://picsum.photos/id/22/800",
-        "https://picsum.photos/id/23/800",
-        "https://picsum.photos/id/24/800",
-        "https://picsum.photos/id/25/800",
-      ],
-    ),
-    ImageCategory(
-      name: 'Teknoloji',
-      icon: Icons.devices,
-      images: [
-        "https://picsum.photos/id/26/800",
-        "https://picsum.photos/id/27/800",
-        "https://picsum.photos/id/28/800",
-        "https://picsum.photos/id/29/800",
-        "https://picsum.photos/id/30/800",
       ],
     ),
   ];
 
-  @override
-  void initState() {
-    super.initState();
-    _tabController = TabController(length: _categories.length, vsync: this);
-  }
+  List<ImageCategory> _userAlbums = [];
+  ImageCategory? _selectedAlbum;
+
+  List<ImageCategory> get _allAlbums => [..._categories, ..._userAlbums];
 
   @override
   void dispose() {
-    _tabController.dispose();
     _searchController.dispose();
     super.dispose();
   }
 
-  void _toggleViewMode() {
-    setState(() {
-      _isListView = !_isListView;
-    });
-  }
-
-  void _changeCrossAxisCount(int count) {
-    setState(() {
-      _currentCrossAxisCount = count;
-    });
-  }
-
-  void _toggleSearch() {
-    setState(() {
-      _isSearching = !_isSearching;
-      if (!_isSearching) {
-        _searchController.clear();
-      }
-    });
-  }
+  ImageCategory get _currentAlbum => _selectedAlbum ?? _categories[0];
 
   List<String> _filterImages(List<String> images) {
     if (!_isSearching || _searchController.text.isEmpty) {
@@ -151,150 +110,215 @@ class _GaleriScreenState extends State<GaleriScreen>
     }).toList();
   }
 
+  String _getImageId(String imageUrl) {
+    final idMatch = RegExp(r'id/(\d+)/').firstMatch(imageUrl);
+    return idMatch?.group(1) ?? 'Bilinmiyor';
+  }
+
+  void _showCreateAlbumDialog() {
+    String albumName = '';
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('Yeni Albüm Oluştur'),
+          content: TextField(
+            decoration: const InputDecoration(hintText: 'Albüm adı'),
+            onChanged: (value) => albumName = value,
+          ),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(context),
+              child: const Text('İptal'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                if (albumName.trim().isNotEmpty) {
+                  setState(() {
+                    _userAlbums.add(ImageCategory(
+                      name: albumName,
+                      icon: Icons.photo_album,
+                      images: [],
+                      isUserCreated: true,
+                    ));
+                    _selectedAlbum = _userAlbums.last;
+                  });
+                }
+                Navigator.pop(context);
+              },
+              child: const Text('Oluştur'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  void _showAddImageDialog() {
+    if (_currentAlbum.isUserCreated == false) return;
+    String imageUrl = '';
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text('Fotoğraf Ekle'),
+          content: TextField(
+            decoration: const InputDecoration(hintText: 'Fotoğraf URL\'si'),
+            onChanged: (value) => imageUrl = value,
+          ),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(context),
+              child: const Text('İptal'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                if (imageUrl.trim().isNotEmpty) {
+                  setState(() {
+                    _currentAlbum.images.add(imageUrl);
+                  });
+                }
+                Navigator.pop(context);
+              },
+              child: const Text('Ekle'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+    final theme = Theme.of(context);
+    final isDark = theme.brightness == Brightness.dark;
+
+    final filteredImages = _filterImages(_currentAlbum.images);
 
     return Scaffold(
-      appBar: AppBar(
-        title: _isSearching
-            ? TextField(
-                controller: _searchController,
-                decoration: InputDecoration(
-                  hintText: 'Fotoğraf ID\'sine göre ara...',
-                  border: InputBorder.none,
-                  hintStyle: TextStyle(
-                    color: isDarkMode ? Colors.white70 : Colors.black54,
+      body: SafeArea(
+        child: Column(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(16.0),
+              color: isDark ? AppColors.darkPrimary : AppColors.primary,
+              child: Row(
+                children: [
+                  _isSearching
+                      ? Expanded(
+                          child: Container(
+                            padding:
+                                const EdgeInsets.symmetric(horizontal: 16.0),
+                            decoration: BoxDecoration(
+                              color: isDark
+                                  ? Colors.grey[800]
+                                  : Colors.white.withOpacity(0.8),
+                              borderRadius: BorderRadius.circular(12),
+                            ),
+                            child: TextField(
+                              controller: _searchController,
+                              decoration: const InputDecoration(
+                                hintText: 'Fotoğraf ID\'sine göre ara...',
+                                border: InputBorder.none,
+                              ),
+                              onChanged: (_) => setState(() {}),
+                              autofocus: true,
+                            ),
+                          ),
+                        )
+                      : Expanded(
+                          child: DropdownButtonHideUnderline(
+                            child: DropdownButton<ImageCategory>(
+                              value: _currentAlbum,
+                              dropdownColor:
+                                  isDark ? Colors.grey[800] : Colors.white,
+                              style: const TextStyle(
+                                  color: Colors.white, fontSize: 16),
+                              iconEnabledColor: Colors.white,
+                              onChanged: (newAlbum) =>
+                                  setState(() => _selectedAlbum = newAlbum),
+                              items: _allAlbums
+                                  .map((album) => DropdownMenuItem(
+                                        value: album,
+                                        child: Text(album.name),
+                                      ))
+                                  .toList(),
+                            ),
+                          ),
+                        ),
+                  IconButton(
+                    icon: Icon(_isSearching ? Icons.close : Icons.search,
+                        color: Colors.white),
+                    onPressed: () =>
+                        setState(() => _isSearching = !_isSearching),
                   ),
-                ),
-                style: TextStyle(
-                  color: isDarkMode ? Colors.white : Colors.black,
-                ),
-                onChanged: (_) => setState(() {}),
-                autofocus: true,
-              )
-            : const Text("Fotoğraf Galerisi"),
-        actions: [
-          IconButton(
-            icon: Icon(_isSearching ? Icons.close : Icons.search),
-            onPressed: _toggleSearch,
-            tooltip: _isSearching ? 'Aramayı Kapat' : 'Ara',
+                  IconButton(
+                    icon: Icon(_isListView ? Icons.grid_view : Icons.view_list,
+                        color: Colors.white),
+                    onPressed: () => setState(() => _isListView = !_isListView),
+                  ),
+                ],
+              ),
+            ),
+            Expanded(
+              child: filteredImages.isEmpty
+                  ? const Center(child: Text('Arama sonucu bulunamadı'))
+                  : Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: _isListView
+                          ? _buildListView(filteredImages)
+                          : _buildGridView(filteredImages),
+                    ),
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          FloatingActionButton(
+            heroTag: "addAlbum",
+            onPressed: _showCreateAlbumDialog,
+            child: const Icon(Icons.create_new_folder),
           ),
-          IconButton(
-            icon: Icon(_isListView ? Icons.grid_view : Icons.view_list),
-            onPressed: _toggleViewMode,
-            tooltip: _isListView ? 'Izgara Görünümü' : 'Liste Görünümü',
-          ),
-          if (!_isListView)
-            PopupMenuButton<int>(
-              tooltip: 'Izgara Boyutunu Değiştir',
-              icon: const Icon(Icons.grid_3x3),
-              onSelected: _changeCrossAxisCount,
-              itemBuilder: (context) => [
-                const PopupMenuItem<int>(
-                  value: 2,
-                  child: Text('2 Sütun'),
-                ),
-                const PopupMenuItem<int>(
-                  value: 3,
-                  child: Text('3 Sütun'),
-                ),
-                const PopupMenuItem<int>(
-                  value: 4,
-                  child: Text('4 Sütun'),
-                ),
-              ],
+          const SizedBox(height: 10),
+          if (_currentAlbum.isUserCreated)
+            FloatingActionButton(
+              heroTag: "addPhoto",
+              onPressed: _showAddImageDialog,
+              child: const Icon(Icons.add),
             ),
         ],
-        bottom: TabBar(
-          controller: _tabController,
-          isScrollable: true,
-          tabs: _categories.map((category) {
-            return Tab(
-              text: category.name,
-              icon: Icon(category.icon),
-            );
-          }).toList(),
-        ),
-        elevation: 4,
-      ),
-      body: TabBarView(
-        controller: _tabController,
-        children: _categories.map((category) {
-          final filteredImages = _filterImages(category.images);
-
-          return filteredImages.isEmpty
-              ? Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Icon(Icons.image_not_supported,
-                          size: 64, color: Colors.grey),
-                      const SizedBox(height: 16),
-                      Text(
-                        'Arama sonucu bulunamadı',
-                        style: Theme.of(context).textTheme.titleLarge,
-                      ),
-                    ],
-                  ),
-                )
-              : Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: _isListView
-                      ? ListView.builder(
-                          itemCount: filteredImages.length,
-                          itemBuilder: (context, index) {
-                            return _buildListItem(
-                                context, filteredImages, index, category);
-                          },
-                        )
-                      : GridView.builder(
-                          gridDelegate:
-                              SliverGridDelegateWithFixedCrossAxisCount(
-                            crossAxisCount: _currentCrossAxisCount,
-                            crossAxisSpacing: 8,
-                            mainAxisSpacing: 8,
-                            childAspectRatio: 1.0,
-                          ),
-                          itemCount: filteredImages.length,
-                          itemBuilder: (context, index) {
-                            return _buildGridItem(
-                                context, filteredImages, index, category);
-                          },
-                        ),
-                );
-        }).toList(),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // Başka bir özellik eklenebilir (örn: favori fotoğrafları gösterme)
-          ScaffoldMessenger.of(context).showSnackBar(
-              const SnackBar(content: Text('Favoriler özelliği yakında!')));
-        },
-        tooltip: 'Favoriler',
-        child: const Icon(Icons.favorite),
       ),
     );
   }
 
-  Widget _buildGridItem(BuildContext context, List<String> images, int index,
-      ImageCategory category) {
-    final String imageUrl = images[index];
+  Widget _buildGridView(List<String> images) {
+    return GridView.builder(
+      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+        crossAxisCount: _gridColumns,
+        crossAxisSpacing: 8,
+        mainAxisSpacing: 8,
+      ),
+      itemCount: images.length,
+      itemBuilder: (context, index) => _buildGridItem(context, images, index),
+    );
+  }
+
+  Widget _buildListView(List<String> images) {
+    return ListView.builder(
+      itemCount: images.length,
+      itemBuilder: (context, index) => _buildListItem(context, images, index),
+    );
+  }
+
+  Widget _buildGridItem(BuildContext context, List<String> images, int index) {
+    final imageUrl = images[index];
+    final theme = Theme.of(context);
+    final isDark = theme.brightness == Brightness.dark;
 
     return GestureDetector(
-      onTap: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => BuyukResimScreen(
-              resimUrl: imageUrl,
-              resimIndex: index,
-              tumResimler: images,
-              kategoriAdi: category.name,
-            ),
-          ),
-        );
-      },
+      onTap: () => _openDetailScreen(context, imageUrl, index, images),
       child: Hero(
         tag: imageUrl,
         child: Container(
@@ -302,7 +326,7 @@ class _GaleriScreenState extends State<GaleriScreen>
             borderRadius: BorderRadius.circular(12),
             boxShadow: [
               BoxShadow(
-                color: Colors.black.withOpacity(0.15),
+                color: isDark ? Colors.black26 : Colors.black12,
                 blurRadius: 5,
                 offset: const Offset(0, 2),
               )
@@ -310,54 +334,21 @@ class _GaleriScreenState extends State<GaleriScreen>
           ),
           child: ClipRRect(
             borderRadius: BorderRadius.circular(12),
-            child: Stack(
-              fit: StackFit.expand,
-              children: [
-                CachedNetworkImage(
-                  imageUrl: imageUrl,
-                  fit: BoxFit.cover,
-                  placeholder: (context, url) => Container(
-                    color: Colors.grey[300],
-                    child: const Center(
-                      child: CircularProgressIndicator(),
-                    ),
-                  ),
-                  errorWidget: (context, url, error) => Container(
-                    color: Colors.grey[300],
-                    child: const Icon(Icons.error),
-                  ),
-                ),
-                Positioned(
-                  bottom: 0,
-                  left: 0,
-                  right: 0,
-                  child: Container(
-                    padding: const EdgeInsets.symmetric(
-                      vertical: 4,
-                      horizontal: 8,
-                    ),
-                    decoration: BoxDecoration(
-                      gradient: LinearGradient(
-                        begin: Alignment.topCenter,
-                        end: Alignment.bottomCenter,
-                        colors: [
-                          Colors.transparent,
-                          Colors.black.withOpacity(0.7),
-                        ],
-                      ),
-                    ),
-                    child: Text(
-                      'ID: ${_getImageId(imageUrl)}',
-                      style: const TextStyle(
-                        color: Colors.white,
-                        fontSize: 12,
-                        fontWeight: FontWeight.bold,
-                      ),
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                ),
-              ],
+            child: CachedNetworkImage(
+              imageUrl: imageUrl,
+              fit: BoxFit.cover,
+              placeholder: (context, url) => Container(
+                color: isDark
+                    ? Colors.grey[800]
+                    : AppColors.secondary.withOpacity(0.5),
+                child: const Center(child: CircularProgressIndicator()),
+              ),
+              errorWidget: (context, url, error) => Container(
+                color: isDark
+                    ? Colors.grey[800]
+                    : AppColors.secondary.withOpacity(0.5),
+                child: const Icon(Icons.error),
+              ),
             ),
           ),
         ),
@@ -365,527 +356,140 @@ class _GaleriScreenState extends State<GaleriScreen>
     );
   }
 
-  Widget _buildListItem(BuildContext context, List<String> images, int index,
-      ImageCategory category) {
-    final String imageUrl = images[index];
+  Widget _buildListItem(BuildContext context, List<String> images, int index) {
+    final imageUrl = images[index];
+    final theme = Theme.of(context);
+    final isDark = theme.brightness == Brightness.dark;
 
     return Card(
       margin: const EdgeInsets.symmetric(vertical: 8.0),
-      elevation: 3,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
+      child: ListTile(
+        leading: Hero(
+          tag: imageUrl,
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(8),
+            child: SizedBox(
+              width: 60,
+              height: 60,
+              child: CachedNetworkImage(
+                imageUrl: imageUrl,
+                fit: BoxFit.cover,
+                placeholder: (context, url) => Container(
+                  color: isDark
+                      ? Colors.grey[800]
+                      : AppColors.secondary.withOpacity(0.5),
+                  child: const Center(child: CircularProgressIndicator()),
+                ),
+              ),
+            ),
+          ),
+        ),
+        title: Text('Fotoğraf ID: ${_getImageId(imageUrl)}'),
+        subtitle: Text('Kategori: ${_currentAlbum.name}'),
+        onTap: () => _openDetailScreen(context, imageUrl, index, images),
       ),
-      child: InkWell(
-        onTap: () {
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => BuyukResimScreen(
-                resimUrl: imageUrl,
-                resimIndex: index,
-                tumResimler: images,
-                kategoriAdi: category.name,
-              ),
-            ),
-          );
-        },
-        borderRadius: BorderRadius.circular(12),
-        child: Row(
-          children: [
-            Hero(
-              tag: imageUrl,
-              child: ClipRRect(
-                borderRadius: const BorderRadius.only(
-                  topLeft: Radius.circular(12),
-                  bottomLeft: Radius.circular(12),
-                ),
-                child: SizedBox(
-                  width: 100,
-                  height: 100,
-                  child: CachedNetworkImage(
-                    imageUrl: imageUrl,
-                    fit: BoxFit.cover,
-                    placeholder: (context, url) => Container(
-                      color: Colors.grey[300],
-                      child: const Center(
-                        child: CircularProgressIndicator(),
-                      ),
-                    ),
-                    errorWidget: (context, url, error) => Container(
-                      color: Colors.grey[300],
-                      child: const Icon(Icons.error),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-            Expanded(
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'Fotoğraf ID: ${_getImageId(imageUrl)}',
-                      style: const TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 16,
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    Text(
-                      'Kategori: ${category.name}',
-                      style: TextStyle(
-                        color: Theme.of(context).colorScheme.secondary,
-                      ),
-                    ),
-                    const SizedBox(height: 4),
-                    Row(
-                      children: [
-                        const Icon(Icons.high_quality, size: 16),
-                        const SizedBox(width: 4),
-                        Text('800 x 800'),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            IconButton(
-              icon: const Icon(Icons.more_vert),
-              onPressed: () {
-                _showImageOptions(context, imageUrl);
-              },
-            ),
-          ],
+    );
+  }
+
+  void _openDetailScreen(
+      BuildContext context, String imageUrl, int index, List<String> images) {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => DetailScreen(
+          imageUrl: imageUrl,
+          index: index,
+          images: images,
         ),
       ),
     );
   }
-
-  String _getImageId(String imageUrl) {
-    final idMatch = RegExp(r'id/(\d+)/').firstMatch(imageUrl);
-    return idMatch?.group(1) ?? 'Bilinmiyor';
-  }
-
-  void _showImageOptions(BuildContext context, String imageUrl) {
-    showModalBottomSheet(
-      context: context,
-      shape: const RoundedRectangleBorder(
-        borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
-      ),
-      builder: (context) {
-        return SafeArea(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              ListTile(
-                leading: const Icon(Icons.favorite_border),
-                title: const Text('Favorilere Ekle'),
-                onTap: () {
-                  Navigator.pop(context);
-                  ScaffoldMessenger.of(context).showSnackBar(
-                      const SnackBar(content: Text('Favorilere eklendi!')));
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.share),
-                title: const Text('Paylaş'),
-                onTap: () {
-                  Navigator.pop(context);
-                  ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-                      content: Text('Paylaşım özelliği yakında!')));
-                },
-              ),
-              ListTile(
-                leading: const Icon(Icons.info_outline),
-                title: const Text('Fotoğraf Bilgisi'),
-                onTap: () {
-                  Navigator.pop(context);
-                  _showImageInfo(context, imageUrl);
-                },
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  void _showImageInfo(BuildContext context, String imageUrl) {
-    showDialog(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: const Text('Fotoğraf Bilgisi'),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text('ID: ${_getImageId(imageUrl)}'),
-              const SizedBox(height: 8),
-              const Text('Boyut: 800 x 800 piksel'),
-              const SizedBox(height: 8),
-              const Text('Format: JPEG'),
-              const SizedBox(height: 8),
-              Text('URL: $imageUrl'),
-            ],
-          ),
-          actions: [
-            TextButton(
-              onPressed: () {
-                Navigator.pop(context);
-              },
-              child: const Text('Kapat'),
-            ),
-          ],
-        );
-      },
-    );
-  }
 }
 
-// Büyük resmi göstermek için geliştirilmiş ekran
-class BuyukResimScreen extends StatefulWidget {
-  final String resimUrl;
-  final int resimIndex;
-  final List<String> tumResimler;
-  final String kategoriAdi;
+class DetailScreen extends StatefulWidget {
+  final String imageUrl;
+  final int index;
+  final List<String> images;
 
-  const BuyukResimScreen({
+  const DetailScreen({
     super.key,
-    required this.resimUrl,
-    required this.resimIndex,
-    required this.tumResimler,
-    required this.kategoriAdi,
+    required this.imageUrl,
+    required this.index,
+    required this.images,
   });
 
   @override
-  State<BuyukResimScreen> createState() => _BuyukResimScreenState();
+  State<DetailScreen> createState() => _DetailScreenState();
 }
 
-class _BuyukResimScreenState extends State<BuyukResimScreen>
-    with SingleTickerProviderStateMixin {
+class _DetailScreenState extends State<DetailScreen> {
   late PageController _pageController;
   late int _currentIndex;
   bool _isFullScreen = false;
-  bool _isInfoVisible = true;
-  bool _isFavorite = false;
-
-  // Basit bir animasyon kontrolcüsü
-  late AnimationController _animationController;
-  late Animation<double> _opacityAnimation;
 
   @override
   void initState() {
     super.initState();
-    _currentIndex = widget.resimIndex;
+    _currentIndex = widget.index;
     _pageController = PageController(initialPage: _currentIndex);
-
-    _animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 300),
-    );
-
-    _opacityAnimation = Tween<double>(begin: 1.0, end: 0.0).animate(
-      CurvedAnimation(
-        parent: _animationController,
-        curve: Curves.easeInOut,
-      ),
-    );
-
-    if (_isFullScreen) {
-      _animationController.forward();
-    }
   }
 
   @override
   void dispose() {
     _pageController.dispose();
-    _animationController.dispose();
     super.dispose();
-  }
-
-  void _toggleFullScreen() {
-    setState(() {
-      _isFullScreen = !_isFullScreen;
-      _isInfoVisible = !_isFullScreen;
-
-      if (_isFullScreen) {
-        _animationController.forward();
-      } else {
-        _animationController.reverse();
-      }
-    });
-  }
-
-  void _toggleFavorite() {
-    setState(() {
-      _isFavorite = !_isFavorite;
-    });
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(
-            _isFavorite ? 'Favorilere eklendi!' : 'Favorilerden çıkarıldı!'),
-        duration: const Duration(seconds: 1),
-      ),
-    );
-  }
-
-  String _getImageId(String imageUrl) {
-    final idMatch = RegExp(r'id/(\d+)/').firstMatch(imageUrl);
-    return idMatch?.group(1) ?? 'Bilinmiyor';
   }
 
   @override
   Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
-    final currentImage = widget.tumResimler[_currentIndex];
-
     return Scaffold(
       extendBodyBehindAppBar: true,
       appBar: _isFullScreen
           ? null
           : AppBar(
-              title: Text(
-                  "${widget.kategoriAdi} - ${_currentIndex + 1}/${widget.tumResimler.length}"),
-              backgroundColor: Colors.black.withOpacity(0.7),
-              elevation: 0,
+              backgroundColor: Colors.black.withOpacity(0.5),
+              title: Text('${_currentIndex + 1}/${widget.images.length}'),
               actions: [
                 IconButton(
-                  icon: Icon(
-                      _isFavorite ? Icons.favorite : Icons.favorite_border),
-                  onPressed: _toggleFavorite,
-                  tooltip: 'Favorilere Ekle/Çıkar',
-                ),
-                IconButton(
-                  icon: const Icon(Icons.share),
-                  onPressed: () {
-                    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-                        content: Text('Paylaşım özelliği yakında!')));
-                  },
-                  tooltip: 'Paylaş',
-                ),
-                IconButton(
                   icon: const Icon(Icons.fullscreen),
-                  onPressed: _toggleFullScreen,
-                  tooltip: 'Tam Ekran',
+                  onPressed: () =>
+                      setState(() => _isFullScreen = !_isFullScreen),
                 ),
               ],
             ),
-      body: Stack(
-        children: [
-          // Ana içerik - PageView
-          GestureDetector(
-            onTap: _toggleFullScreen,
-            child: Container(
-              color: Colors.black,
-              child: PageView.builder(
-                controller: _pageController,
-                itemCount: widget.tumResimler.length,
-                onPageChanged: (index) {
-                  setState(() {
-                    _currentIndex = index;
-                  });
-                },
-                itemBuilder: (context, index) {
-                  return InteractiveViewer(
-                    minScale: 0.5,
-                    maxScale: 4.0,
-                    child: Center(
-                      child: Hero(
-                        tag: widget.tumResimler[index],
-                        child: CachedNetworkImage(
-                          imageUrl: widget.tumResimler[index],
-                          fit: BoxFit.contain,
-                          placeholder: (context, url) => const Center(
-                            child:
-                                CircularProgressIndicator(color: Colors.white),
-                          ),
-                          errorWidget: (context, url, error) => Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              const Icon(Icons.error,
-                                  color: Colors.white, size: 48),
-                              const SizedBox(height: 8),
-                              Text('Resim yüklenemedi: $error',
-                                  style: const TextStyle(color: Colors.white)),
-                            ],
-                          ),
-                        ),
+      body: GestureDetector(
+        onTap: () => setState(() => _isFullScreen = !_isFullScreen),
+        child: Container(
+          color: Colors.black,
+          child: PageView.builder(
+            controller: _pageController,
+            itemCount: widget.images.length,
+            onPageChanged: (index) => setState(() => _currentIndex = index),
+            itemBuilder: (context, index) {
+              return InteractiveViewer(
+                minScale: 0.5,
+                maxScale: 4.0,
+                child: Center(
+                  child: Hero(
+                    tag: widget.images[index],
+                    child: CachedNetworkImage(
+                      imageUrl: widget.images[index],
+                      fit: BoxFit.contain,
+                      placeholder: (context, url) => const Center(
+                        child: CircularProgressIndicator(),
                       ),
-                    ),
-                  );
-                },
-              ),
-            ),
-          ),
-
-          // Resim bilgisi kartı - Animasyonlu göster/gizle
-          if (_isInfoVisible)
-            Positioned(
-              top: MediaQuery.of(context).padding.top +
-                  (AppBar().preferredSize.height * 1.1),
-              right: 16,
-              child: FadeTransition(
-                opacity: _opacityAnimation,
-                child: Card(
-                  color: (isDarkMode
-                          ? const Color.fromARGB(255, 34, 33, 33)
-                          : Colors.white)
-                      .withOpacity(0.8),
-                  elevation: 4,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(12),
-                  ),
-                  child: Padding(
-                    padding: const EdgeInsets.all(12.0),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          'Fotoğraf Bilgisi',
-                          style: TextStyle(
-                            fontWeight: FontWeight.bold,
-                            fontSize: 16,
-                            color: isDarkMode ? Colors.white : Colors.black87,
-                          ),
-                        ),
-                        const SizedBox(height: 8),
-                        Text(
-                          'ID: ${_getImageId(currentImage)}',
-                          style: TextStyle(
-                            color: isDarkMode ? Colors.white70 : Colors.black54,
-                          ),
-                        ),
-                        const SizedBox(height: 4),
-                        const Text(
-                          'Boyut: 800 x 800',
-                          style: TextStyle(
-                            color: Colors.grey,
-                          ),
-                        ),
-                        const SizedBox(height: 4),
-                        Text(
-                          'Kategori: ${widget.kategoriAdi}',
-                          style: TextStyle(
-                            color: isDarkMode ? Colors.white70 : Colors.black54,
-                          ),
-                        ),
-                      ],
+                      errorWidget: (context, url, error) => const Icon(
+                        Icons.error,
+                        size: 48,
+                      ),
                     ),
                   ),
                 ),
-              ),
-            ),
-
-          // Kaydırma indikatörü
-          Positioned(
-            top: MediaQuery.of(context).size.height / 2,
-            left: 8,
-            child: FadeTransition(
-              opacity: _opacityAnimation,
-              child: _currentIndex > 0
-                  ? IconButton(
-                      icon: Container(
-                        decoration: BoxDecoration(
-                          color: Colors.black45,
-                          shape: BoxShape.circle,
-                        ),
-                        child: const Padding(
-                          padding: EdgeInsets.all(4.0),
-                          child:
-                              Icon(Icons.arrow_back_ios, color: Colors.white),
-                        ),
-                      ),
-                      onPressed: () {
-                        _pageController.previousPage(
-                          duration: const Duration(milliseconds: 300),
-                          curve: Curves.easeInOut,
-                        );
-                      },
-                    )
-                  : const SizedBox.shrink(),
-            ),
+              );
+            },
           ),
-
-          Positioned(
-            top: MediaQuery.of(context).size.height / 2,
-            right: 8,
-            child: FadeTransition(
-              opacity: _opacityAnimation,
-              child: _currentIndex < widget.tumResimler.length - 1
-                  ? IconButton(
-                      icon: Container(
-                        decoration: BoxDecoration(
-                          color: Colors.black45,
-                          shape: BoxShape.circle,
-                        ),
-                        child: const Padding(
-                          padding: EdgeInsets.all(4.0),
-                          child: Icon(Icons.arrow_forward_ios,
-                              color: Colors.white),
-                        ),
-                      ),
-                      onPressed: () {
-                        _pageController.nextPage(
-                          duration: const Duration(milliseconds: 300),
-                          curve: Curves.easeInOut,
-                        );
-                      },
-                    )
-                  : const SizedBox.shrink(),
-            ),
-          ),
-        ],
+        ),
       ),
-      bottomNavigationBar: _isFullScreen
-          ? null
-          : BottomAppBar(
-              color: Colors.black.withOpacity(0.8),
-              child: Padding(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    IconButton(
-                      icon:
-                          const Icon(Icons.arrow_back_ios, color: Colors.white),
-                      onPressed: _currentIndex > 0
-                          ? () {
-                              _pageController.previousPage(
-                                duration: const Duration(milliseconds: 300),
-                                curve: Curves.easeInOut,
-                              );
-                            }
-                          : null,
-                      color: _currentIndex > 0 ? Colors.white : Colors.white38,
-                    ),
-                    Text(
-                      '${_currentIndex + 1} / ${widget.tumResimler.length}',
-                      style: const TextStyle(color: Colors.white, fontSize: 16),
-                    ),
-                    IconButton(
-                      icon: const Icon(Icons.arrow_forward_ios,
-                          color: Colors.white),
-                      onPressed: _currentIndex < widget.tumResimler.length - 1
-                          ? () {
-                              _pageController.nextPage(
-                                duration: const Duration(milliseconds: 300),
-                                curve: Curves.easeInOut,
-                              );
-                            }
-                          : null,
-                      color: _currentIndex < widget.tumResimler.length - 1
-                          ? Colors.white
-                          : Colors.white38,
-                    ),
-                  ],
-                ),
-              ),
-            ),
     );
   }
 }
