@@ -46,7 +46,7 @@ class WelcomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
+      backgroundColor: const Color(0xFFF5F2E8), // Updated background color
       body: SafeArea(
         child: Padding(
           padding: const EdgeInsets.all(24.0),
@@ -194,7 +194,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
+      backgroundColor: const Color(0xFFF5F2E8), // Updated background color
       appBar: AppBar(
         title: const Text('Kayıt Ol'),
         backgroundColor: Colors.transparent,
@@ -350,7 +350,7 @@ class _LoginScreenState extends State<LoginScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
+      backgroundColor: const Color(0xFFF5F2E8), // Updated background color
       appBar: AppBar(
         title: const Text('Giriş Yap'),
         backgroundColor: Colors.transparent,
